Using Requirement.swift instead of custom model
