--- Sources/TuistCore/Models/DependencyRequirement.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-enum DependencyRequirement {
-    case exact(String)
-}
